rename styles base -> todoScreen
